+++ packages/docs/src/config.ts
@@ -18,10 +18,11 @@ const conf = defineConfig(
 				utils,
 				opts : { emojis: { 'utils-media': '📷' } },
 			} )
-			const pkg     = typeof utils?.config?.const?.pkg === 'object' ? utils.config.const.pkg : {}
-			const data    = await docs.getPkgConfig( pkg )
-			const input   = resolve( '../../docs' )
-			const output  = resolve( './build' )
+
+			const pkg    = typeof utils?.config?.const?.pkg === 'object' ? utils.config.const.pkg : {}
+			const data   = await docs.getPkgConfig( pkg )
+			const input  = resolve( '../../docs' )
+			const output = resolve( './build' )
 
 			return {
 				...data,