+++ packages/config/stylelint-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dovenv/stylelint-config
 
+## 1.5.26
+
+### Patch Changes
+
+- Update predocs
+
 ## 1.5.25
 
 ### Patch Changes