+++ packages/create/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-dovenv
 
+## 1.5.26
+
+### Patch Changes
+
+- Update predocs
+
 ## 1.5.25
 
 ### Patch Changes